requirements.txt
flask==2.1.3
requests==2.28.1
pillow==9.2.0
